QR書き換え用/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
autopep8==2.3.1
black==24.10.0
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
dill==0.3.9
exceptiongroup==1.2.2
fastapi==0.88.0
flake8==7.1.1
Flask==3.0.3
Flask-Session==0.8.0
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.5.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
macholib==1.16.3
MarkupSafe==3.0.2
mccabe==0.7.0
modulegraph==0.19.6
msgspec==0.18.6
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
py2app==0.28.8
pycodestyle==2.12.1
pydantic==1.10.18
pydantic_core==2.23.4
pyflakes==3.2.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pylint==3.3.1
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.22.0
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.4
zipp==3.20.2
